Revert "Some ultra dirty hacks to reoslve YT and CH logging macro clashes."
This reverts commit 15bd83dd8d1bfb8ce7dd8206def3f561b8d2dc41.

+++ yt/server/clickhouse_server/engine/config_repository.cpp
@@ -7,10 +7,10 @@
 
 #include <yt/server/clickhouse_server/native/storage.h>
 
-//#include <Poco/Logger.h>
-//#include <Poco/Util/XMLConfiguration.h>
+#include <Poco/Logger.h>
+#include <Poco/Util/XMLConfiguration.h>
 
-//#include <common/logger_useful.h>
+#include <common/logger_useful.h>
 
 #include <util/string/cast.h>
 
@@ -121,8 +121,8 @@ NNative::TObjectAttributes TConfigRepository::GetAttributes(const std::string& n
 
 bool TConfigRepository::LooksLikeConfig(const NNative::TObjectAttributes& attributes) const
 {
-    return static_cast<int>(attributes.Type) == static_cast<int>(NNative::EObjectType::Document) ||
-           static_cast<int>(attributes.Type) == static_cast<int>(NNative::EObjectType::File);
+    return attributes.Type == NNative::EObjectType::Document ||
+           attributes.Type == NNative::EObjectType::File;
 }
 
 IConfigPollerPtr TConfigRepository::CreatePoller(const std::string& name) const
